- Replaced many sendf by sendbin (less allocation) - Removed many strlen - /!\ Header size is now hardcoded

+++ src/json.h
@@ -47,7 +47,7 @@ struct json_childs {
 
 struct jsontring {
 	char *jstring;
-	size_t jsize;
+	int jsize;
 };
 
 void set_json(char *name, char *value, struct json **jprev);
